Extract `toc/snapshots.rs`
